Refactor Compiler to allow for arbitrary binary operators
Components of the compiler are now split into different Isabelle
theories. This has also allowed us to abstract away a plus, or etc to
a binary operator to make codegen and type checking easier.

Types for parsing and types for compiling ASTs have also been unified
to make life easier.