Add support for let bindings
This adds support in the grammar, compiler and type system for these
bindings. Understanding memory layout was particularly difficult. It
appears that memory is byte-addressed, even though we can store 256-bit
words in there. In the future, I should find a more elegant way to index
them.

I had to make changes to the parser (add more rule-depth) to avoid
shift-reduce conflicts in let bindings. I opted for semicolon syntax
rather than "in" syntax for familiarity, but the semantics are the
same.