Add function definitions to language
This only adds definitions, not calling. A single function can
be parsed and compiled, but never executed. The execution of the
main body of a program is left unaffected.

Extensive use of records has been applied, as there is a lot of
information that needs to be passed around the compiler. This 'program context'
contains function definitions and the main expression, currently.

Parsing and pretty printing is working.