Enable deployment of contracts on test blockchain
This involves being able to call the compiler from our JS test-suite.
This was straight-forward enough, we simply needed to make sure there
was no garbage output. What took the longest time was realising that
newlines in the compiler output completely frazzled the EVM.

We also had to write initialisation code for deployed contracts. What is
deployed to Ethereum is actually a bytecode snippet that when executed
returns the bytecode to run on every execution of the function. This was
easy enough with an instruction that copies program code into memory.