Add support for external function calls
This adds the ability for you to call an exported function
off-chain (and on-chain), pass in arguments and get a result.

This is done using the Solidity ABI, so we've had to work to convert
Solidity's ideas to NewLanguage's. Namely, a function with return type
Unit actually has "no outputs" in Solidity parlance (similarly for
inputs).

The grammar and type system have been amended to allow for "export fun".
The type of a program is now the types of the function it exports (along
with its name).