Add record types
This adds support for records in NewLanguage. Records are represented
with {  } and the contents inbetween. Types use colons and assignments
use =. Unit has been replaced with {  }.

This allows functions to have multiple inputs and outputs, a very useful
feature. In particular for external facing functions. These results are
abstracted to the Solidity ABI.

Tests have been added and the grammar amended.