Add persistent state, reading and side effects
Side effects have been added to the type system, grammar and codegen.
Codegen is largely unaffected, as side effects are modelled as an effects system.
Functions can have effects, and calling them, and then unwrapping them,
brings the effects into the current scope.

This is modelled using a TEffect type. This has caused difficulty in
numerous areas, as TEffect is often a special case.
For example, in the ABI, the signature of a TEffect is just the signature
of its inner type.

Persistent state is first created with an init function, which contracts
must have. This returns the initial state of the contract.
To read from state, you can wrap a function that accepts state as an
argument with a with_state modifier, which provides the state and
creates an effectful function. This allows us to enforce that reading
or writing to state can only be done once
for any given execution. This is important for re-entrancy protection.

Work still needs to be done to enforce this is the case.

Interesting bugs cropped up on the way in implementing this. Debugging
was extensive, including binary diffing (note for dissertation).