Add persistent state modification
We augment the available function modifiers to include "updating_state"

This takes a function which takes in some state and an extra argument,
and returns a new state and some return value. The augmented function
simply takes the argument, and returns the return value, changing the
state in the process (with the appropriate effect types)

In writing tests for this, it was important to remember the difference
between .call and .send for these methods. .call simply evaluates the
function while .send actually issues a transaction. Important when
persisting state!