Add record updates
Records are now stored similar to vtables in OOP languages.

This allows them to inherit from other records, so updates can be
performed efficiently. This is especially the case when it comes to
state. Unecessary reads of state do not need to be performed, and when
writing updates back, only the values that have changed can be saved.