Add ability to send Ether
This adds the "send" expression to the language, which is an effectful
expression that sends ether to a given address. Included is a new type
for addresses, because modelling addresses as integers will allow for
arithmetic and that's not good.

Also included are the protections against unchecked send and
re-entrancy, which were straight-forward. Tests are added to check for
this, which includes setting up a re-entrancy attack, verifying it works
on a vulnerable contract, and then verifying the attack does not work on
the contract written in the new language.