Fix moving of state
The optimisation that only stored updated values in state assumed that
if a record eventually pointed to storage, it didn't need to be saved.
This is false, as we may be moving an existing record in storage to a
new location. The optimisation has been updated to check that if a
record based on storage is being saved, it only avoids updating if it's
being saved in the same place as the record it is based on.