Add mapping types
Mapping types allow for relating scalar data (ints, bools and addresses)
to other kinds of data.

This adds support for them by adding them to the grammar, type system
and codegen. Mappings are created with explicit types (to prevent the
need for type inference on empty mappings), and come with an update
syntax analogous to records.

Storing mappings in state has been non-trivial. The
save_state_at_address function has assumed we know the address at
compile time. As mappings have arbitrary keys, this is no longer true.
Getting around this has been tedious.

Secondly, the type system allows from moving a mapping from one place in
state to another. For example, from one record field to another. This is
currently impossible with the implementation of mappings, so the
contract will crash at runtime if this is attempted. This could be
improved, by detecting it at the type checking stage or changing the
implementation so such moves are actually possible.