Optimise binary size
This is done by trimming down most words to their smallest
representation in their instructions. In some areas, this was
problematic, such as in function calls. These change the size of the
function's instructions which mean that we don't know where to call.

The only way to reasonably fix this is to use a proper assembler with
labels to jump to.

A little hack was employed to determine if an address was in memory or
state. A 1 in the leftmost bit is a negative number, so instead we can
do a comparison to 0.