Add balance, value and current contract address to language
I've faked the existence of records that match these values by having
this.address and this.balance be tokens that parse to the required AST
nodes. Long term, not a great solution but it proves the point.

This has also fixed some long-standing issues with ABI generation. State
mutability is now reflected correctly. We can no longer send money to
contracts that don't want them.